Change DelaunayTriangulation and AlphaShape to use stl
This changes the DelaunayTriangulation and AlphaShape
classes to use std::vector's and std::pair's instead of
QList's, QVector's, and QPair's.

+++ src/chemkit/coordinates.cpp
@@ -300,7 +300,7 @@ Point3 Coordinates::center() const
 /// position with \p weights.
 Point3 Coordinates::weightedCenter(const std::vector<Float> &weights) const
 {
-    Q_ASSERT(size() == weights.size());
+    Q_ASSERT(static_cast<unsigned long>(size()) == weights.size());
 
     if(isEmpty()){
         return Point3();
